requirements.txt
anthropic==0.21.3
openai==1.12.0
PyYAML==6.0.1
requests==2.31.0
streamlit==1.32.2
gradio==4.24.0
python-dotenv==1.0.1
requests==2.31.0
biopython>=1.83
pdfplumber
pandas
python-docx